🧩 📦 Refactor, Struct: 1 mv, 1 rn.
